fix: Reorganizing for a better file structure and mantainment
