+++ ecommerce-app/index.ts
@@ -1,56 +1,16 @@
-import { ApolloServer, gql } from "apollo-server";
+import { ApolloServer } from "apollo-server";
 import { IResolvers } from "@graphql-tools/utils";
-import * as data from './mockdata.json';
 
-const typeDefs = gql`
-    type Product {
-        id: ID!
-        name: String!
-        description: String!
-        quantity: Int!
-        price: Float!
-        onSale: Boolean!
-        image: String!
-        category: Category!
-    }
-    type Category {
-        id: ID!
-        name: String!
-        products: [Product!]!
-    }
-    type Query {
-        products: [Product!]!
-        product(id: ID!): Product
-        categories: [Category!]!
-        category(id: ID!): Category
-    }
-`;
+import { typeDefs } from "./schema";
+
+import { Query } from "./resolvers/Query";
+import { Product } from "./resolvers/Product";
+import { Category } from "./resolvers/Category";
+
 const resolvers: IResolvers = {
-    Query: {
-        products: () => data.products,
-        product: (parent,args, context) => {
-            return data.products.find(product => product.id === args.id);
-        },
-        categories: (parent, args, context) => {
-            return data.categories; 
-        },
-        category: (parent, args, context) => {
-            const { id } = args;
-            return data.categories.find(category => category.id === id);
-        }   
-    },
-    Category: {
-        products: (parent, args, context) => {
-            const { categoryId } = parent;
-            return data.products.filter(product => product.categoryId === categoryId);
-        }
-    },
-    Product: {
-        category: (parent, args, context) => {
-            const { categoryId } = parent;
-            return data.categories.find(category => category.id === categoryId);    
-        }
-    },
+    Query,
+    Product,
+    Category,
 };
 
 const server = new ApolloServer({
