+++ ecommerce-app/resolvers/Product.ts
@@ -1,0 +1,14 @@
+import { IResolvers } from "@graphql-tools/utils";
+
+import * as data from '../mockdata.json';
+
+const Product: IResolvers  = {
+    category: (parent, args, context) => {
+        const { categoryId } = parent;
+        return data.categories.find(category => category.id === categoryId);    
+    }
+}
+
+export {
+    Product
+}
